Update PlaneHandler.java
Sorry for intrusion, Sasha. But I have interested in your linkedin profile and then github profile. :) Want to suggest you some changes. Maybe you would like them ;O)

"Deciding whether to use if-then-else statements or a switch statement is based on readability and the expression that the statement is testing. An if-then-else statement can test expressions based on ranges of values or conditions, whereas a switch statement tests expressions based only on a single integer, enumerated value, or String object. " Which is your case I would say. And using enumerations, you can define and reuse that string constants.

Also about line #26 & 48. Not sure if it is needed, if qName could be nullable. Or maybe you need lombok.NonNull annotation to add to parameter in the method in order to raise exception. It is up to you.

Sorry, I did not test code. Just did modification here in github editor. 

Also it is a bit strange prefix 'temp' (tempValue, tempPlane, etc.) It reminds me some temporary variable used in local 'for' loop when you switch two vars. Prefix 'temp' is not suitable to instance variables, I think.

Please, don't be angry on me ;) You are doing great job! I just want to share my view.